requirements.txt
fastapi==0.75.2
numpy==1.22.3
onnxruntime-gpu==1.11.0
opencv-python==4.5.5.64
Pillow==9.1.0
